+++ server/src/test/java/org/realityforge/gwt/appcache/server/TestPropertyProvider.java
@@ -3,13 +3,13 @@ package org.realityforge.gwt.appcache.server;
 import javax.servlet.http.HttpServletRequest;
 import org.realityforge.gwt.appcache.server.propertyprovider.PropertyProvider;
 
-final class TestPropertyProvider
+public final class TestPropertyProvider
   implements PropertyProvider
 {
   private final String _key;
   private final String _value;
 
-  TestPropertyProvider( final String key, final String value )
+  public TestPropertyProvider( final String key, final String value )
   {
     _key = key;
     _value = value;