+++ server/src/test/java/org/realityforge/gwt/appcache/server/mgwt/MgwtManifestServlet.java
@@ -1,0 +1,14 @@
+package org.realityforge.gwt.appcache.server.mgwt;
+
+import org.realityforge.gwt.appcache.server.propertyprovider.UserAgentPropertyProvider;
+
+public class MgwtManifestServlet
+  extends AbstractMgwtManifestServlet
+{
+  public MgwtManifestServlet()
+  {
+    addPropertyProvider( new MgwtOsPropertyProvider() );
+    addPropertyProvider( new UserAgentPropertyProvider() );
+    addPropertyProvider( new MobileUserAgentProvider() );
+  }
+}